separate Maze and Sphere classes into dedicated files
